Update for WWDC25. Update GitHub username in README.
WWDC25 moved to a single page driven by JavaScript, the video data
coming from a single JSON source.

WwdcJsonParser handles most of the work loading and interpreting
the JSON, deferring to InfoExtractor for handling m3u8 data for
each video.

Video data is not parsed until a topic's videos are requested to
avoid unnecessarily downloading and processing m3u8 data for videos
that will not be downloaded.